Added owl-carousel in landing page.
- Hard coded images for carousel.
