+++ .gitignore
@@ -8,6 +8,8 @@
 # dependencies
 /node_modules
 
+/build
+
 # profiling files
 chrome-profiler-events.json
 speed-measure-plugin.json
