+++ package.json
@@ -27,6 +27,7 @@
     "font-awesome": "^4.7.0",
     "hammerjs": "^2.0.8",
     "moment": "^2.24.0",
+    "ngx-owl-carousel-o": "^1.0.11",
     "rxjs": "~6.3.3",
     "tslib": "^1.9.0",
     "zone.js": "~0.8.26"
