--- src/app/home/home.component.html
@@ -1,3 +1,0 @@
-<p>
-  home works!
-</p>
