+++ src/app/landing/landing.component.html
@@ -1,0 +1,13 @@
+<div class="container">
+  <div>Some tags before</div>
+  <owl-carousel-o [options]="customOptions">
+    <ng-template carouselSlide><img src="../../assets/image/dummy/mike-anderson-66252-unsplash.jpg" /></ng-template>
+    <ng-template carouselSlide><img src="../../assets/image/dummy/diego-jimenez-258120-unsplash.jpg" /></ng-template>
+    <ng-template carouselSlide><img src="../../assets/image/dummy/ian-dooley-407846-unsplash.jpg" /></ng-template>
+    <ng-template carouselSlide><img src="../../assets/image/dummy/alexander-london-390806-unsplash.jpg" /></ng-template>
+    <ng-template carouselSlide><img src="../../assets/image/dummy/mike-anderson-66252-unsplash.jpg" /></ng-template>
+    <ng-template carouselSlide><img src="../../assets/image/dummy/ian-dooley-407846-unsplash.jpg" /></ng-template>
+    <ng-template carouselSlide><img src="../../assets/image/dummy/diego-jimenez-258120-unsplash.jpg" /></ng-template>
+  </owl-carousel-o>
+  <div>Some tags after</div>
+</div>
