+++ src/app/landing/search/search.component.html
@@ -1,0 +1,9 @@
+<form class="site-search" method="get">
+  <input type="text" name="site_search" placeholder="Type to search...">
+  <div class="search-tools">
+    <span class="clear-search">Clear</span>
+    <span class="close-search">
+      <i class="icon-cross"></i>
+    </span>
+  </div>
+</form>
